Dynamic buffer size in transaction log writer
Change transaction log writer buffer size depending on the number of
available CPUs.
The default buffer size is 512KB as it was before for small machines.
Runtimes with a higher number of CPUs will get bigger buffer size up to 4MB.
Also, switch to using direct buffer instead of using on-heap buffer.