Handle pruning var expand with predicates in slotted
For queries like `()-[r* {p:42}]->()` where we have one or more predicates
associated with the variable expand was not running on the slotted runtime
since we never allocated slots for the variables introduced in the predicates.

There was also a second bug that would have hit us if we had supported these
types of queries since in the interpreted fallback we were using
`ExecutionContext::remove` which is only implemented in
`MapExecutionContext`.