Clear interruption status in TimeoutRetrierTest
TimeoutRetrier catches InterruptedException and re-interrupts the
thread. This behavior was tested but the interruption status was
not cleared. Thread went on to execute other tests and could fail
when blocking operation was executed. The very first blocking
operation would fail because the thread remained interrupted.
Often the first test to fail was PanicServiceTest because it is
sometimes executed directly after TimeoutRetrierTest.

Problem is fixed by clearing the interruption status
in TimeoutRetrierTest.