Fix flaky RotatableCsvOutputIT
Metric type was changed and that changed a number of columns in result CSV file.
As result, csv file rotation appeared before
expected moment because of the byte size rotations criteria used in the test.
That was causing infinite looping while waiting for a line from the
wrong file.
Also, update most of the metrics tests to junit 5.