Add last checkpointed tx to StoreCopyFinshedResponse
If response was unsuccessful or from an earlier protocol than value will
be -1.

Add a TxPullRequestExecutor

It is a stateful executor which will retry tx
pulling until either successful or maximum consecutive failures has
been reached.

A failure is defined by either an error response or no progress
response.

The executor will route requests depending on its state. All normal
requests will be sent to "secondary()" address. The "primary()" address
is only used if:

- This is the last try after failing consecutively
- Or, the pull reuqest is considered complete because re constraint has
been met and a Success response has been received. One last request
is then sent to the "primary()" address, this last request is allowed
to fail.