Add logging around folder changes in backup
Logging is for the case when the target backup directory exists and
is not empty. Backup is then performed into a temporary directory
like 'graph.db.temp.0'. Then pre-existing directory is moved to a
temporary directory like 'graph.db.err.0'. And temporary backup
directory 'graph.db.temp.0' is moved to the specified target path.