Improve SubProcess dispatcher wait
The code regarding the dispatcher looks a bit racy and might be the cause
of a flaky test on OS X, so I've improved it slightly. If the intention here
is that the observation of the process having finished is a trigger for
ensuring visibility of and returning the dispatcher, then there is no need
to return a raw null, but rather whatever the value of the dispatcher is
should be used.

The code for catching the IllegalThreadStateException is also part
of Java 8 now, so that was replaced with Process#isAlive().