Make the token stores cope with name record cycle inconsistencies.
This is important because we will now read the tokens on possibly inconsistent stores in tools like the Consistency Checker.
The TokenStore will just ignore tokens that have cycle inconsinstencies.
To avoid this becomming a drag on performance, we only enable this check for the TokenStore, because that is the only place where it is strictly needed.
All other places should, during normal operation, ideally never see such cycles, so we disable the check in those cases.