Remove the trace logging from the fulltext index accessor and populator.
This logging was put in place in the hopes of catching the cause of a flaky test.
However, the test has not failed for nearly 3 months, though it used to fail at a much higher frequency than that.
Let's assume that the cause, whatever it might have been, has been fixed, and remove the logging.
If the flakiness returns, we can put the logging back.