Restore all previously cleaned up high limit formats and tests.
Because of the previously incorrect selection of the latest high limit format
for store upgrade it was possible that high limit stores with old version
were not upgraded even if upgrade was forced. Now with fixed selection
we need to be able to migrate previously non upgraded stores and to be
able to do that we need to restore all old high limit formats.