Uses scope:runtime on most neo4j-record-storage-engine
At least on most components using this dependency. This is to further
limit leakage of its details into the rest of the product.

Making this restriction highlighted abstraction problems with
store migration, which have been split into storage migration
and database migration, living in kernel and storage respectively.

Hiding the commands also highlighted the need to expose token ids
from token commands so created StorageCommand.TokenCommand
which has an accessor for its tokenId(). This is used at least
in causal-clustering.