Separate errors between transient and not transient in tx pulling.
The TxPulRequestExecutor uses a timeout based approach to when it gives
up. However some errors are less likely to ever be resolved with retries.

If an error is considered not transient will now immediately move state
to LAST_ATTEMPT.

Errors considered transient are:
ConnectException, CatchupResolutionException

Errors not considered transient are:
All E_* responses and any exception not mentioned above.

There are quite likely more exceptions that could be considered
transient.