Lock down and fix tests around database file locations
The exact location of files is a not very easy to work with
concoction of configuration, assumptions, hardcoded behaviour etc.

This change locks down the configuration inherited and passed
on by the bootstrapper and fixates the location of files in the
test to be well alinged with the common setup of files in an actual
installation.

A better future would be one where the entire configuration was not
passed through at all, but rather just the particular pieces as
required per component. Unfortunately, this style of parameter passing
is used all the way to the bottom layers, f.ex. LogFilesBuilder.