Fixed many CC backup/restore ITs
 - Addedd OverriddenDatabaseConfig allows the user
to specify a number of config values which differ from the global
config. These overrides are *not* augments on the global config, and are
scoped to the specificDatabaseConfig.

- Several backup/restore tests were not working, first because they were
creating system dbs but not backing them up/restoring them then second
because of the lack of overridden config caused record format selection
problems. These tests now backup and restore system dbs properly, as
well as graph.db