Ability to drop UNIQUE index using DROP INDEX query
The type comparison became problematic in the scenario where a
uniqueness index for a uniqueness constraint was created,
but the constraint creation got aborted or crashed, leaving the
unique index still in existence. Trying to drop this index using
DROP INDEX cypher construct would not succeed because
the index which was put into the transaction state would always
be an index of type "general" and would not be recognized upon commit.