CompositeGenericKey don't use PageCursor.getOffset when reading keys
Why was this a problem?
When reading keys using CompositeGenericKey, PageCursor.getOffset() was
used to calculate how large the remaining key should be after reading only
part of it.
When reading keys from a block in BlockStorage, ReadableChannelPageCursor
is used. This is a helper class, presenting a ReadableChannel as a read-only
PageCursor. The problem is that there is no way for ReadableChannelPageCursor
to implement getOffset method because ReadableChannel by itself does not
know about position and the implementations that do, such as ReadAheadChannel
use long for position. PageCursor.getOffset() must return int.
This made it impossible for BlockStorage to deal with CompositeGenericKey.

Why is it safe?
When reading a composite key using CompositeGenericKey the keySize is only
used as a best effort safety net. It sets an upper bound on how large a
key containing a string can be. Using the full keySize, without decrementing
it as we read more and more parts of the composite key, is good enough for
this purpose.
It also acts as a safety net, guarding that we don't try and read any key
with size so small that it doesn't even fit the TYPE_ID (indicating what
type the key has (integer, long, string, etc). This guard is also best
effort.