Ensure that token create commands update the token holders during forward-recovery.
This is necessary because e.g. fulltext indexes need to read from the token holders when they create index proxies.
This also means that we have to properly reverse index create commands during reverse recovery, since the indexing service
would otherwise see indexes that are "created in the future" and try to create index proxies for them, and this could be
happening before the token create commands have been applied, and then we would run into the same problem with the
fulltext indexes.