Pass comparator into SortPipe, PartialSortPipe
Just as it was done already for TopPipe, we pass the comparator
into SortPipe and PartialSortPipe. The advantage is that we don't need
slotted versions, since they would now do exactly the same. All slot
accesses happen from the comparator. This will make slotted partial
sort faster.

As a side effect, this enables us to remove SortSlottedPipe.