Batching and chunking for transactions pulled by catchup protocol
TxPullResponses are now batched and/or chunked to fixed size buffers.
For tiny transactions this will reduce overhead since we previously
would send every tx in its own buffer together with some metadata. For
larger transactions we avoid the risk of slow allocation if exceeding
nettys pooled buffer chunk size.

Most of the chuning is handled by the encoder but there are a few
changes in how the response messages are send.

1. ResponseMessageType TX is only sent once in the beginning of the tx
stream - Before this would be sent before every tx.

2. After all tx have been written to the pipeline a final empty
TxPullResponse is written so that the encoder knows it should flush and
complete the final chunk.
3. Store Id is only included in the first chunk of a tx stream - before
this was sent together with each tx.