(Re)allow read replicas to store copy from multiple addresses
This functionality was removed due to vulnerabilities in the previous
catchup protocol. V3 fixes this and since we will remove support for
earlier protocols in 4.0 we can safely do this again.

This will vastly improve store copy robustness in Read Replicas since,
given the selection strategy allows, we can switch instance in the
middle of store copy and we do not have to retry from start.