Re-introduce OnlineBackup API
This API is used to take backups from a client application. It
existed in the previous branch and relied on HA/com backup
functionality. It was later removed together with HA.

New version of the API uses catchup protocol, same as the backup
command line tool.

Changes in OnlineBackup compared to the previous version:

 * removed deprecated methods for performing full/incremental backup

 * added fluent interface methods for configuring backup settings
   and dependencies (like output stream) instead of method parameters

 * keep a single method to execute the actual backup as
   a terminal operation

 * return a Result object from OnlineBackup#backup() method instead
   of modifying the OnlineBackup instance with the consistency check
   status of the last backup

 * added null-checks for method parameters