Fix error response handling in catchup V3
Netty pipelines for client and server did not have correct handlers
installed and were not able to encode/decode error responses. This
resulted in protocol violations and generic failures with
ClosedChannelException. Client pipeline was missing an error
response handler. Server pipeline was missing an error response
encoder.

This commit fixes the problem by adding the missing handlers
to client and server pipelines. The change is tested by ITs that
request files for an unknown database name. Such requests
are expected to either return the correct status code or
throw an exception with good error message.