Remove --name argument from backup
This argument defined a sub-directory of the --backup-dir where
database files will get placed. Example:

```
$neo4j-admin backup --backup-dir=/var/backups --name=myBackup --database=system.db
```

resulted in the following directory structure:

```
/var/backups/myBackup/
├── neostore.nodestore.db
├── neostore.relationshipstore.db
└── ...
```

so database name was not present in the final directory tree. This
caused various confusions when restoring the database because kernel
currently treats directory name as the database name.

Having a mandatory --name argument and optional --database argument
does not make sense in a multi-database setup. Both --backup-dir
and --name define two parts of the same thing - target path. In
multi-database setup --database argument will eventually become
mandatory.

This commit removes the --name argument and make backup create a
directory with the same name as the database name. Example:

```
$neo4j-admin backup --backup-dir=/var/backups --database=system.db
```

results in the following directory structure:

```
/var/backups/system.db/
├── neostore.nodestore.db
├── neostore.relationshipstore.db
└── ...
```