Backport: Retry dropping index on DirectoryNotEmptyException or NoSuchFileException
On index drop we recursively delete the index directory. If a PageCache
profile file is created simultaneously there is a risk that we run into
DirectoryNotEmptyException. In that case, we just retry the operation
again.

Adding NoSuchFileException to list of acceptable exceptions because it
only means that the file we where trying to delete was deleted by someone
else, which should be fine.