Handle queries with PROFILE in reactive API
If you ran a write query with PROFILE where there were no rows to be
returned in commercial, the query failed because the call to
`onCompleted(queryStatistics())` was being called after the transaction
was closed. This caused an issue since in commercial we expect to be
able to fetch page cache counters from the transaction when collecting
statistics, resulting in a null pointer exception for this case.