Moves LabelScanStore out from kernel to new component
There's one code change to this which is relevant to bring up
and that is regarding the LabelScan interface. It used to return
IndexProgressor instances and accept IndexProgressor client.
Since IndexProgressor is not quite a kernel-api thing, but just
an implementation detail if that api inside kernel then obviously
LabelScan couldn't continue looking like that. It has been changed
to return PrimitiveLongResourceIterator and kernel will instead
wrap those in an IndexProgressor. This should be monitored for
performance difference, but suspected that it will be negligable.
The code is also simpler because the LabelScanValueProgressor
doesn't need to duplicate the code to read and return ids from
LabelScanValue instances.