Multidatabase: database selection support in Bolt
With multidatabse, Bolt server needs to select the correct database to start a transaction based on a database name passed at the beginning of a transaction (a.k.a. via either `Session#run` or `Session#beginTx`).
Bolt server perform database selection by creating the correct `TransactionStateMachine` towards the correct database.
With single database, each Bolt connection initialize `TransactionStateMachine` after Init/Hello message and then hold on to this `TransactionStateMachine` for the whole life time of the connection.
With multidatabase, Bolt need to decouple `TransactionStateMachine` from connection, moving the creation of `TxStateMachine` to the point where database name is available.
That is when we are in the `READY` state and then we received a `RUN` message or `BEGIN` message.
The `TransactionStateMachine` will then be kept inside the connection's state until a `RESET` message arrives to clear the reference to it.

This current implementation relies on the `RESET` message to release resources towards the database.