TransactionStateMachine release after commit
After a transaction is closed (committed successfully or rolled back with an error), the `TransactionStateMachine` is no longer needed. Instead of relying on `RESET` message to clear the resources, we can immediately release the `TransactionStateMachine`. This is achieved by a callback from `TransactionStateMachine` to `BoltStateMachine` to let `TransactionStateMachine` unregister/release itself from `BoltStateMachine` once a transaction is done.