Fix failing tests caused by PULL_ALL/DISCARD_ALL message after COMMIT/ROLLBACK in BoltV1 and BoltV2
As we are release TransactionStateMachine after tx close (such as COMMIT/ROLLBACK), the PULL_ALL/DISCARD_ALL errors when trying to streaming result via non-existing TransactionStateMachine.
Thus we moved the logic of streaming COMMIT/ROLLBACK as well as BEGIN into `Ready` state in BoltV1. Removes the logic of an extra dummy result in TransactionStateMachine.