Make Raft server creation independent from core components
Raft server is going to be created and started only once because it
is a global component. Thus its creation should be independent of
database-specific components.

This commit turns RaftServerModule into a RaftServerFactory which
is independent from core components like CoreLife and
CoreDownloadFactory. These components are now added to the global
lifecycle in core edition module.