Select neo4j or graph.db as a default database.
Based on the presence of legacy graph.db and configuration of legacy
`active_database` or new `default database` setting we will select
`graph.db` or new `neo4j` as default database.
By default `neo4j` is default database.
in case if legacy `graph.db` is present and `default database`
setting is not configured to point to some database
`graph.db` will be selected as default.
In case if legacy `active_database` is specified its value be migrated
and assigned to `default database` setting.