Refactor cluster state to better support multi database
Introduced ClusterStateLayout that encapsulates logic for global
and per-database files/directories needed in cluster state. It is
meant to the only component that knows about file/directory names
used to store the state.

Removed code that performed migrations of old `cluster-state`
directories. Made lifecycle management of every storage explicit
instead of internal to the component that creates it. Removed
a cache of created storages.