Refactor Zombie - upgrade argumentState to be explicit buffer...
The argumentStateMap is now treated as a buffer in the same way as
MorselBuffer and MorselArgumentBuffer. This means that pipelines can
put their output morsels directly into the ArgumentStateBuffer, which
group the morsel rows by argumentRowId, and updates the relevant
accumulators. The ArgumentStateBuffer is also responsible for only
allowing pipelines to take accumulators that are have seen all the
relevant argument rows.

This change removes the MorselExecutionContext.counters, which felt like
a strange place to have reference counting. It also renames lot's of
abstractions to less confusing names.