More efficient schema cache/lookup in IndexMap
Introduces a sort of multi-set which combines label and property keys
so that a more narrow view of related indexes can be selected.
For CREATE/DELETE of nodes the selection can be made 100% precise
since all the data is there (the previous lookup method wasn't precise),
but for UPDATE where node data is incomplete, more specifically
the property keys, the selection isn't as precise.