Remove unused mocking from RelationshipRecordFormatTest.
It turns out that the unused mock was actually making the test flaky, because the mock could be garbage collected (since it was never used) before the stubber had finished instrumenting it.
It then also turns out that Mockito has checks for this, and will throw an exception from the stubber if it detects that the mock (which Mockito itself is only weakly holding on to) has been garbage collected.