Fix SegmentedRaftLogVerificationIT on Windows
The test was changed to use real file system instead of an ephemeral
one. This made it fail on Windows because created Raft logs were
never closed and TestDirectory extension was not able to cleanup
the directory.

This commit fixes the problem by making the test always
shutdown the created Raft log.