More flexible AllStoreHolder#indexReference
It used to throw ISE on a very specific case where it wanted
to convert an IndexDescriptor --> IndexReference for a committed
index that couldn't be looked up from IndexingService.
Now it instead instantiates a StoreIndexDescriptor and continues
about its business. The need to do this should be very rare
and seems to only happen with full-text indexes in and around
their creation and/or replication in clustering.
The test that occasionally ran into this is
FulltextIndexCausalClusterIT.