Add optional database name parameter to routing procedure
This new parameter defines for which database the routing
table should be returned. Default database is used when
parameter is absent. Error is thrown when the specified
database does not exist. Single instance verifies existence
of the database using DatabaseManager. Cores and read replicas
will throw when discovery service does not contain any
information about the specified database.

Updated procedure signature is:

```
dbms.routing.getRoutingTable(context :: MAP, database = null :: STRING)
```

Filtering by database name on cores and read replicas is
missing because discovery service contains multi-cluster
database names. Filtering added but commented out with
TODOs and can be uncommented once multi-database names
are stored in the discovery service. This is also the
reason for a disabled test in ClusterDiscoveryIT.