Uses efficient schema matching on constraints checking
IndexMap caches constraints in addition to indexes. This allows for
more efficient schema matching based on the change, instead of
going through all constraints and filter.

The reading was further cleaned up by making a property key list
read in one place and pass in array instead of doing that read
arbitrarily in multiple places. This makes the code clearer
in what it does and prevents duplicate work.