More resilient maxMemory calculation in importer
Importer calculates available memory by getting total machine memory
minus max heap memory and multiply that by some factor.
This works well most of the time, but if max heap memory is
configured very high, the import performance will suffer,
very likely in vain because not much heap is needed,
but much off-heap memory. It can even be the case that a heap memory
which is larger than machine memory is configured, which results
in available memory for the importer to be negative.

To accommodate for this the calculation will cap the max heap memory
figure at half machine memory. In very rare cases for heavily
restrained environments this could become a problem, but overall
this change should solve more problems than it introduces.