Fix download of snapshot for several databases
This is a temporary workaround which is prudent to have while we work
on separating out databases into separate Raft groups. It caused
some errors in stress tests when the neo4j database and an associated
snapshot was downloaded before the system database and its snapshot,
because the last snapshot could make the Raft skip ahead of the previous
store, which meant that you could get a Raft that is ahead of the store
with unapplied changes, effectively missing commands and causing corruption.