Use local variables to represent slots
Within a fused operator loop we map execution context slots
to local variables as needed, and avoid reading/writing intermediate
values to the ExecutionContext. At the innermost level we either
write these local variables to the execution context,
or just directly to the result in case we have a ProduceResult.