Use configured root path in database factories
Right now database factories take the parent of the specified
directory as the databases root path. This is done for
backward compatibility with previous single-database code.

Before this change, database factories always re-configured the
databases root path even when it has already been provided in
the config. This made it hard to start a database exactly in
the specified folder.

This commit makes community and commercial database factories use
the configured databases root path.