BlockBasedIndexPopulator can handle entries up to ~8k
Most significant bit in second byte in DynamicSizeUtil can either be
used as an indicator that the entry exist on offload or it can be
used as part of key size to support key sizes up to ~8k.

BlockBasedIndexPopulator (BlockEntry) use the bit as part of key size
because it doesn't have any offload store nor does it need one.

Add test for GenericBlockBasedIndexPopulator to make sure it handles
large values and fail on too large values.