Extract bootstrapping of individual databases in CC
Refactor core bootstrapping to have a component that bootstraps
the given database. Before this change, all databases were
bootstrapped together. This is a step towards multi-raft where we
need to bootstrap databases separately and at runtime.

Changed store initialization to run in a temporary directory
`temp-bootstrap` under the database directory. This makes it
possible to start a separate temporary database that does not
interfere with the real database. Setting to ignore the store
lock was removed because of this.

Bootstrapping of existing stores without ID files uses recovery
to re-create them. Recovery checker updated to consider stores
with missing ID files and such that require recovery. All
databases are recovered before bootstrapping in CoreLife.

All snapshots produced by the bootstrap procedure still contain
shared RaftCoreState and GlobalSessionTrackerState because all
databases live in the same Raft group. This will change once
we go multi-raft.