Always initialize nested lifecycle in TransactionLogsRecovery
TransactionLogsRecovery manages lifecycle of the given schemaLife.
Initialization of TransactionLogsRecovery can short-circuit but
should always initialize schemaLife. Otherwise, it is both
initialized and started when TransactionLogsRecovery is
started which results in an incorrect recovery lifecycle.

The problem became visible since recovery check started saying
recovery is required for stores with missing ID files.