Replace CopyOnWriteHashMap with ConcurrentHashMap
This commit removes CopyOnWriteHashMap and makes code use the
standard ConcurrentHashMap instead. Custom implementation of
the concurrent map breaks easily when default methods are
added to the Map interface. Compiler does not force CopyOnWriteHashMap
to implement any of them and we silently have not thread safe
or not atomic as the result.

For example, method #computeIfAbsent() was inherited from the
Map interface and implemented without any synchronization. So
it could end up invoking the given re-mapping function more
then once if multiple threads were invoking this method for
the same key.

Thus, it is easier to remove the custom concurrent map
implementation than to override all default methods after
every new Java release.

Also, fixed a non-atomic compiler cache in Cypher codebase.