Lucene index migration.
The `lucene-1.0`, `lucene+native-1.0`, and `lucene+native-2.0` index providers are now gone.
The oldest providers are migrated to the `lucene+native-3.0` provider, and the `lucene+native-2.0` provider is migrated to the `native-btree-1.0` provider.
This includes modifying index schema rules in the schema store during store migration, and deleting the provider root directories of the old providers.
This means that we now only ever store _string values_ in Lucene, and the now-unused support for other value types have been removed from ValueEncoding and LuceneDocumentStructure.