OutputOperator -
This new type of operator occurs only as the last operator of a pipeline,
and performs the work of putting the output of the task into the correct
downstream sink. The OutputOperator implementers should include regular
morsel output and produce results, but also PreSort and PreAggregation.

Execution happens as a 2-step process, where the output is
first prepared, and the produced.

1. prepare()
Perform transformations and pre-computations on the output morsel that are
possible without synchronization with other threads. This step will be
performed as part of regular PipelineTask execution. The point is to
minimize the amount of work done under lock.

2. produce()
Put the prepared output into the correct sink/consumer. This step will
introduce contention, and is considered part of scheduling.