Simplify capturing of the kernel commit process in CC
Created a commit process factory that creates both simple and
replicated commit processes. First one is directly installed
into state machines. Second one is given to kernel. This
change simplifies the previous commit installation that
required a listener and relied on lifecycle methods.

Disabled ClusteringLocalOperatorIT because it does not work
at the moment. Databases can't be created and started
locally because every database in CC needs to be bootstrapped
and have at least a magic transaction log entry.