Flat map database factories and builders
Since the introduction of the system database, we now have more than one database which the API should reflect. The new top-level API is `DatabaseManagementService` that will expose the multiple `GraphDatabaseService`s. The goal is also to start with a builder and get a service, hiding the factory from the public API giving us more flexibility when changing the factory.

* `GraphDatabaseFactory` is now `DatabaseManagementServiceBuilder`

* `GraphDatabaseFacadeFactory` is now `DatabaseManagementServiceFactory`

* `GraphDatabaseBuilder` is merged into `DatabaseManagementServiceBuilder`

* `GraphDatabaseBuilder.DatabaseCreator` is removed
The `DatabaseCreator` was a way to defer the creation of the actual database when the builder was done. This backward way of doing things was a relic of the past when you first created a factory, extracted a builder from that, and then got your service.