Change cluster role procedure for multi-db
Before this change, procedure `dbms.cluster.role` did not take
any parameters and always returned the same cluster role. This
was not sufficient in multi-database setup because every
database can be in a separate Raft group and have different role.

This commit adds a database name parameter to the procedure. It is
a mandatory parameter and nulls are not permitted. The returned
role is now role of the cluster member in the Raft group for
the specified database.

Example:

```
CALL dbms.cluster.role("neo4j")

╒═════════╕
│"role"   │
╞═════════╡
│"LEADER" │
└─────────┘
```