More generic exception check in backup stress test
A shutdown database now responds with an error response
(instance of CatchupErrorResponse) on every request. This results
in two different exception types being thrown on the client
side - StoreCopyFailedException and StoreIdDownloadFailedException.
The later is thrown only when attempting to fetch a store ID and
the database is shut down.

Backup stress test performs retries on transient failures. It did
not consider StoreIdDownloadFailedException to be transient
and failed.

This commit makes the transient exception check in backup
stress test more generic.  It is also changed to accept
StoreIdDownloadFailedException.