Simplifications based on kernel code review
Notably the fact that the ThreadToStatementContextBridge is shared by all databases
and so there is no value in thinking the ContextSwitchingSystemGraphQueryExecutor
needs access to each database, it only needs the ThreadToStatementContextBridge.

Also the SystemDatabaseInnerAccessor was created in the context that already contained
the GraphDatabaseService for the SystemDatabase, and as such does not need to look
that up later when needing to create TransactionContext for executing queries.