Fixed an error in bolt where failure message is not flushed when no free work thread is available.
When the worker thread pool is full, bolt server will fail new requests from clients. However this error is not flushed correctly to the client.
This PR make bolt server to first flush the failure message to client and then close the connection to enforce the driver failing fast without waiting for more responses.