Use PatternExpressionSolver for SubQuery plans.
---

Using the solver for `expr` in all SemiApply-like plans is kinda stupid.
The idea is that `expr` is cheap to evaluate while the subquery (`inner`) is costly.
If `expr` is _also_ a PatternExpression or PatternComprehension, that is not true any longer,
and it could be cheaper to execute the one subquery  (`inner`) instead of the other (`expr`).

This is not a regression anyway, since before this change, the `expr` subquery
was also solved, but at runtime.

The above is an observation that actually there should be some cost
estimation if multiple subqueries are involved, whether they are solved
with `RollUpApply` (`PatternExpressionSolver`) or `XXSemiApply` (`selectPatternPredicates`),
to see which is the cheaper subquery and should be executed first.