Version cluster state and drop it when migrating to 4.0
Recreation of the cluster state directory is needed because it
has a different structure in 4.0. Changes in its structure
happened because of multi-database.

Introduced new simple storage in cluster state which contains
its version. Currently, this version is 1 and all cluster state
directories without the version file are considered to be pre 4.0.
Versioning should allow easier migrations/changes to cluster
state in the future.